feat(core): C4 复杂判定入口统一 + clippy -D warnings 清零 + 质量 CI + SSRF 防线加固
- signals::is_complex（complexity_score ≥ 0.5）单一真源，router band_for 委托；
  原 router 私有正则并入 complexity_rx 并集，阈值调优仍待标注语料（台账 C4）
- clippy 1.97 全量 lint 清零（clamp/is_empty/sort_by_key/map_identity 等 37 处）
- 新增 .github/workflows/ci.yml：Rust clippy+test / WebUI tsc+vite / Python
  pyflakes+pyright 三质量门
- SSRF/路径加固：_plan_subtask 环回/内网校验（httpx 化）、embedding 下载仅限
  配置镜像主机（httpx 流式）、对账报告 containment、测试值运行时构造

+++ crates/lloom-core/src/conversations.rs
@@ -73,7 +73,7 @@ fn new_id() -> String {
     let ms = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap_or_default()
-        .as_millis() as u128;
+        .as_millis();
     let pid = std::process::id() as u128;
     format!("{:016x}", (ms << 20) | (pid & 0xf_ffff))
 }
